mm: Introduce Reported pages
In order to pave the way for free page reporting in virtualized
environments we will need a way to get pages out of the free lists and
identify those pages after they have been returned. To accomplish this,
this patch adds the concept of a Reported Buddy, which is essentially
meant to just be the Uptodate flag used in conjunction with the Buddy
page type.

To prevent the reported pages from leaking outside of the buddy lists I
have added a call to clear_reported_page to the del_page_from_free_list
function. As a result any reported page that is split, merged, or
allocated will have the flag cleared prior to the PageBuddy value being
cleared.

The process for reporting pages is fairly simple. Once the nr_free for
a given free area has exceeded the number of pages reported for that area
plus a certain high watermark value we will flag the zone as needing
reporting and schedule the worker thread to start reporting. That worker
thread will begin working from the lowest supported page reporting order
up to MAX_ORDER - 1 pulling unreported pages from the free list and
storing them in the scatterlist.

When processing each individual free list it is necessary for the worker
thread to release the zone lock when it needs to stop and report the full
scatterlist of pages. To reduce the work of the next iteration the worker
thread will rotate the free list so that the first unreported page in the
free list becomes the first entry in the list. Doing this we should only
have to walk an entire free list no more than twice assuming large
numbers of pages are not being added to the tail of the list.

It will then call a reporting function providing information on how many
entries are in the scatterlist. Once the function completes it will return
the pages to the tail of the free area from which they were allocated and
start over pulling more pages from the free areas until there are no
longer enough pages to report on to keep the worker busy.

The worker thread will work in a round-robin fashion making its way
though each zone requesting reporting, and through each reportable free
list within that zone. Once all free areas within the zone is below the
high watermark level for free pages to report the flag indicating that the
zone has requested reporting will be cleared, and if no zones are
requesting reporting the worker thread will exit.

Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>