Introduce page pinning IOMMU driver
The aim of this driver is to provide cooperative enlightenment for
making DMA work with passthrough devices in the presence of memory
overcommt. In hooks into dma_ops and does roughly two things:
- access each page when it's mapped for DMA to ensure it's resident
- maintain a ref count of how many times the page has been mapped for
DMA. The ref counts are shared with the hypervisor so that the
hypervisor knows not to swap DMA-able pages out

The ref count is only partially complete currently; it's not actually
being shared with the hypervisor now so it's essentially useless.
Also it needs to be converted to a flat bitmap.
Once we have a device exposed then this driver can share the bitmap with
the hypervisor, and also use that device for enabling this driver.

Seeing as there is no device now we use a kernel cmd line are to enable
the device:
page-pinning-iommu.enable=y

Signed-off-by: James Gowans <jgowans@amazon.com>