kvm: gfn_to_pfn lookup handles huge PFNMAP PTEs
A PFNMAP VMA may have huge PTEs in the page table. Cater for that by
walking the userspace page tables using follow_invalidate_pte which is
enlightened to huge mappings, and return the size of the mapping found
to the caller via an output argument similar to how we indicate to
callers whether the PFN is writeable or not.

The size output argument is only non-zero for PFNMAP VMAs. This is a
mechanism to indicate to the caller that we know the size of the PFN. In
the case of a non-PFNMAP VMA, size will be zero indicating that the
caller must use existing mechanisms for THP or hugetlbfs to figure out
the size. Perhaps gfn_to_pfn could handle these THP and hugetlbfs cases
in future, but just handing PFNMAP for now.

This should not be a behaviour change to existing callers: currently the
path is
__gfn_to_pfn_memslot
    hva_to_pfn
        hva_to_pfn_remapped
	    follow_pte
At present follow_pte does not return huge PTEs, it will -EINVAL via
follow_invalidate_pte because the pmdpp/pudpp arguments aren't supplied.
Hence the behvior or supplying these arguments when calling from
hva_to_pfn_remapped but returning an error unless the size output
argument is supplied results in the same caller experience when size is
not supplied.

Signed-off-by: James Gowans <jgowans@amazon.com>

CR: https://code.amazon.com/reviews/CR-52925878