mm: use pmd_leaf not pmd_huge in follow_invalidate_pte
The *_huge functions are tied to CONFIG_HUGETLB_PAGE whereas the _leaf
functions are a more portable way of detecting a huge entry, indepentent
of hugetlbfs beig compiled in.
This allows PFNMAP VMAs which huge entries to be detected even when
hugetlbfs is not enabled.

Signed-off-by: James Gowans <jgowans@amazon.com>

CR: https://code.amazon.com/reviews/CR-52925698