mm: Split huge PUD on wp_huge_pud fallback
Currently the implementation will split the PUD when a fallback is taken
inside the create_huge_pud function. This isn't where it should be done:
the splitting should be done in wp_huge_pud, just like it's done for
PMDs. Reason being that if a callback is taken during create, there is
no PUD yet so nothing to split, whereas if a fallback is taken when
encountering a write protection fault there is something to split.

It looks like this was the original intention with the commit where the
splitting was introduced, but somehow it got moved to the wrong place
between v1 and v2 of the patch series. Rebase mistake perhaps.

Fixes: 327e9fd48972 ("mm: Split huge pages on write-notify or COW")

Cc: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Christian König <christian.koenig@amd.com>
Signed-off-by: James Gowans <jgowans@amazon.com>