iommufd: Add plumbing for KHO (de)serialise
To support serialising persistent iommufd objects to KHO, and to be able
to restore the persisted data it is necessary to have a serialise hook
on the KHO active path as well as a deserialise hook on module init.

This commit adds those hooks and the new serialise.c file which will
hold the logic here; for now it's just empty functions.