iommufd: Re-hydrate a usable iommufd ctx from sysfs
When the sysfs file is read, create an iommufd file descriptor, create a
fresh iommufd_ctx, and populate that ictx struct and related structs
with the data about mapped IOVA ranges from KHO.

This is done in a super yucky way by having the sysfs file's .show()
callback create a new file and then print out the new file's fd number.
Done this way because I couldn't figure out how to define a custom
.open() callback on a sysfs object.
An alternative would be to have a new iommufd pseudo-filesystem which
could be mounted somewhere and would have all of the relevant persistent
data in it.

Opinions/ideas on how best to expose persisted domains to userspace are
welcome.